Add support for key schemas
fixes: #301

BREAKING CHANGE

With Kafka, different parts of the message payload, i.e. the Kafka record, can have type and schema information associated with them.  Most people think of the record value, but there is also the key to think about, and in the future the headers.

This change seeings support for providing type & schema information for the key of each Kafka record in a topic.  This is done by requiring the message payload to define both a `key` and `value` section. Hence this is a BREAKING CHANGE.

Previously, the value schema could be supplied via a reference:

```yaml
  message:
    payload:
      $ref: "/schema/value.avsc"
```

To support key schemas this is now achieved by changing the spec to:

```yaml
  message:
    payload:
      key:
        $ref: "/schema/key.avsc"
      value:
        $ref: "/schema/value.avsc"
```

The key, or indeed the value, can also be defined as holding one of the inbuilt Kafka types, i.e. types for which most Kafka clients provide serializers out of the box, e.g.

```yaml
  message:
    payload:
      key:
        type: string
      value:
        $ref: "/schema/value.avsc"
```

Support for inline schemas for both key and value remains, however, the provisioning of inline schemas is still unsupported.