fixes: #301
BREAKING CHANGE

With Kafka, different parts of the message payload, i.e. the Kafka record, can have type and schema information associated with them.  Most people think of the record value, but there is also the key to think about, and in the future the headers.

This change seeings support for providing type & schema information for the key of each Kafka record in a topic.  This is done by extracting the key information from the channel's `message.bindings.kafka.key` field.

For example, the following defines a Kafka topic with a key that contains a complex type:

```yaml
  message:
    bindings:
      kafka:
        key:
          $ref: "/schema/key.avsc"
    payload:
      $ref: "/schema/value.avsc"
```

The key, or indeed the value, can also be defined as holding one of the inbuilt Kafka types, i.e. types for which most Kafka clients provide serializers out of the box, e.g.

```yaml
  message:
    payload:
      key:
        type: string
      value:
        $ref: "/schema/value.avsc"
```

Support for inline schemas for both key and value remains, however, the provisioning of inline schemas is still unsupported.