Enable SASL on internal listener when enabling SASL.
The broker created by the `DockerKafkaEnvironment` test utility has two listeners: one used by other containers within the Docker network (`BROKER`) and one used by test code (`PLAINTEXT`: terrible confusing name!).

When SASL is enabled, previously only the `PLAINTEXT` was being switched to have SASL enabled with `SASL_PLAINTEXT`. The docker network listener, `BROKER`, was left as `PLAINTEXT`.

This mean any other Docker containers, including Schema Registry, would connect on an unauthenticated connection, and assigned the `ANONYMOUS` user: which was added as a super-user.

This is bad!

If people want to test their services and components work and they're running them in Docker, then those components needs to connect to an endpoint that requires authentication and, hence, has ACLs applied.

This PR enables SASL on the docker network listener `BROKER` when SASL is enabled, and only adds the `ANONYMOUS` super-user if SASL is not enabled.