Spec's returned from parser should be valid. (#400)
`operationId` is a required field of `Operation`, but the check for it is only done in the provision call. This means the parser can return a spec that is invalid = bad.

This change makes the parsing fail if the `operationId` is missing.

The change also applies the "don't allow your objects to get into an invalid state" best practice. Which removes the need for a `validate` method.

Co-authored-by: Andy Coates <acoates@marketaxess.com>