Flip default of `schema.reflection` to `false` for new clients
The deprecated `Clients` method set `SCHEMA_REFLECTION_CONFIG` to `true`. This is _not_ a good default for production.

This setting means serde types will use reflection to _build_ a schema from a Pojo. Specmesh is all about taking manual control of the schema. So we shouldn't be using generated schemas.

Additionally, this setting is not need / actually causes problems, when the types in question are generated from the proto / avro schema, e.g. avro specific record types.

We should be encouraging people to generate pojos from the schema, rather than hand-crafting them.

This change flips the setting at the default `false`. It also uses the avro plugin to generate a specific record type from a schema and uses this in some test cases to show it works.

Other functional tests have been left using the hand-crafted Pojo: these need to have reflection turned on.