SchemaReaders now use String, not Path, for resource locations of schema (#568)
On Windows a `Path` uses `\` as the separator, where as a resource path in Java should always use `/`.

Hence, SchemaReaders should not use `Path` for locations that are potentially class path locations.

Deprecated methods that take a `Path` for potential classpath locations and added `String` method variants.

Added a much-needed unit test for `SchemaReaders`.

Co-authored-by: Andy Coates <acoates@marketaxess.com>